refactor: 지메일 스레드 락 획득 로직 공통화 (#34)
