+++ worker/src/main/java/com/mailsangja/worker/service/mail/GmailHistoryStateApplyCommandService.java
@@ -1,6 +1,5 @@
 package com.mailsangja.worker.service.mail;
 
-import com.mailsangja.db.entity.mail.GmailThreadLock;
 import com.mailsangja.db.entity.mail.MailAccount;
 import com.mailsangja.db.entity.mail.Message;
 import com.mailsangja.db.entity.mail.Thread;
@@ -12,20 +11,14 @@ import com.mailsangja.worker.common.exception.mail.MailPushException;
 import com.mailsangja.worker.dto.gmail.GmailHistoryEvent;
 import com.mailsangja.worker.dto.mail.InitialMailSyncThreadSaveCommand;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.exception.ConstraintViolationException;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.sql.SQLException;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class GmailHistoryStateApplyCommandService {
-
-    private static final String GMAIL_THREAD_LOCK_UNIQUE_CONSTRAINT = "uq_gmail_thread_locks_account_thread";
-    private static final String POSTGRES_UNIQUE_VIOLATION_SQL_STATE = "23505";
 
     // INBOUND/OUTBOUND row가 분리되어 있어도 Gmail read state는 동일 thread 기준으로 집계한다.
     private final GmailThreadLockRepositoryPort gmailThreadLockRepositoryPort;
@@ -41,7 +34,7 @@ public class GmailHistoryStateApplyCommandService {
             boolean read,
             InitialMailSyncThreadSaveCommand syncCommand
     ) {
-        acquireThreadLock(mailAccount, event.gmailThreadId());
+        gmailThreadLockRepositoryPort.acquireThreadLock(mailAccount, event.gmailThreadId());
 
         if (syncCommand != null && !gmailHistoryStateQueryService.existsMessage(
                 mailAccount.getId(),
@@ -105,51 +98,7 @@ public class GmailHistoryStateApplyCommandService {
         message.markAsUnread();
     }
 
-    private void acquireThreadLock(MailAccount mailAccount, String gmailThreadId) {
-        gmailThreadLockRepositoryPort.findByMailAccountIdAndGmailThreadIdAndDeletedAtIsNullForUpdate(
-                mailAccount.getId(),
-                gmailThreadId
-        ).orElseGet(() -> createAndLockThreadLock(mailAccount, gmailThreadId));
-    }
-
-    private GmailThreadLock createAndLockThreadLock(MailAccount mailAccount, String gmailThreadId) {
-        try {
-            gmailThreadLockRepositoryPort.save(GmailThreadLock.builder()
-                    .mailAccount(mailAccount)
-                    .gmailThreadId(gmailThreadId)
-                    .build());
-        } catch (DataIntegrityViolationException e) {
-            // 동시에 같은 thread lock row를 생성한 경우를 허용하고, 그 외 제약조건 문제는 다시 실패시킨다.
-            if (!isConcurrentLockCreation(e)) {
-                throw e;
-            }
-        }
-
-        return gmailThreadLockRepositoryPort.findByMailAccountIdAndGmailThreadIdAndDeletedAtIsNullForUpdate(
-                        mailAccount.getId(),
-                        gmailThreadId
-                )
-                .orElseThrow(() -> new MailPushException(MailPushErrorCode.GMAIL_HISTORY_RESULT_INVALID));
-    }
-
     private boolean isBlank(String value) {
         return value == null || value.isBlank();
-    }
-
-    private boolean isConcurrentLockCreation(DataIntegrityViolationException exception) {
-        Throwable cause = exception.getCause();
-        while (cause != null) {
-            if (cause instanceof ConstraintViolationException constraintViolationException
-                    && GMAIL_THREAD_LOCK_UNIQUE_CONSTRAINT.equals(constraintViolationException.getConstraintName())) {
-                return true;
-            }
-
-            if (cause instanceof SQLException sqlException
-                    && POSTGRES_UNIQUE_VIOLATION_SQL_STATE.equals(sqlException.getSQLState())) {
-                return true;
-            }
-            cause = cause.getCause();
-        }
-        return false;
     }
 }
